refactor: BookingQueueRepository -> BookingQueueManager
