+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/booking/service/BookingService.java
@@ -23,7 +23,7 @@ import com.pgms.apibooking.domain.booking.dto.response.BookingGetResponse;
 import com.pgms.apibooking.domain.booking.dto.response.BookingsGetResponse;
 import com.pgms.apibooking.domain.booking.dto.response.PageResponse;
 import com.pgms.apibooking.domain.booking.repository.BookingQuerydslRepository;
-import com.pgms.apibooking.domain.bookingqueue.repository.BookingQueueRepository;
+import com.pgms.apibooking.domain.bookingqueue.service.BookingQueueManager;
 import com.pgms.apibooking.domain.payment.dto.request.PaymentCancelRequest;
 import com.pgms.apibooking.domain.payment.dto.request.RefundAccountRequest;
 import com.pgms.apibooking.domain.payment.service.PaymentService;
@@ -59,7 +59,7 @@ public class BookingService { //TODO: 테스트 코드 작성
 	private final TicketRepository ticketRepository;
 	private final MemberRepository memberRepository;
 	private final BookingQuerydslRepository bookingQuerydslRepository;
-	private final BookingQueueRepository bookingQueueRepository;
+	private final BookingQueueManager bookingQueueManager;
 	private final SeatLockManager seatLockManager;
 	private final PaymentService paymentService;
 	private final TossPaymentConfig tossPaymentConfig;
@@ -166,7 +166,7 @@ public class BookingService { //TODO: 테스트 코드 작성
 
 	@Async
 	protected void removeSessionIdInBookingQueue(Long eventId, String tokenSessionId) {
-		bookingQueueRepository.remove(eventId, tokenSessionId);
+		bookingQueueManager.remove(eventId, tokenSessionId);
 	}
 
 	private EventTime getBookableTimeWithEvent(Long timeId) {
